Bleeding: Let Passable prefer the data's set-back location.
This might not fully fix all fly bypasses, but it prevents the Passable
check from resetting players into the air for moves that may cover any
distance from air into ground, leading to resetting the player into the
air, which allows for a fly bypass, effectively. Preferring the data's
set-back location will set the player back onto ground if it is
passable,
this still leaves some area for cheat attempts, but the order of speed
and passable checks shows the dilemma that either can undo the others
effects with the set-back locations used by the checks. [Subject to
review.]