Fix passable alerts with fence gates.
This workaround allows the block transition from above a 1.5 high block
into the block, for the case both from and to have been inside of the
bounds of the block. This should fix the annoying bug with jumping
inside of an open fence gate, closing it while just above 1.0 height,
so that passable would alert for moving back from air into the block.