"lost sprint": Remove checking vs horizontal buffer.
Might consider for future:
The problems of the type lost-sprint, bunny seem to be client mechanics.
The ground-modifiers apply and stay valid during jump, thus turning off 
sprint for attacking makes sense.

So i revert the statement about "could be worth a ticket" for the
moment. However with looking forward to have a Minecraft API, it could 
could be appropriate to provide something to allow server-side checking,
yet again finding the right ground modifier needs knowledge of the 
exact moves the client did, so client side could change to send all the 
transition events as suggested before.