Add option to assume the player is sprinting when possible (read on).
This is a workaround for the case that sprinting events are missing,
wrongly set up or for events firing in an unusual order.

This does allow speeding hacks that allow players to go at sprinting
speed without telling the server, so it should only be turned on if
there is any issues.

The option has to be turned on: checks.moving.assumesprint