Angle - Allow more configuration
You can now change the threshold per each check instead of one universal threshold. 

Notes for the new thresholds:
averageTime: If the average time elapsed is between 0 and 150 millisecond(s), add it to the violation. (Even more configuration here?)
averageMoves: If the average move is between 0 and 0.2 block(s), add it to the violation.
averageYaw: If the average difference of yaw is superior to 50 degrees, add it to the violation. (More configuration here?)
averageSwitch: No information from comments. Assuming this checks if the player is switching targets too quickly?